+++ src/Control/Natural/IsoF.hs
@@ -72,6 +72,8 @@ isoF
     -> f <~> g
 isoF = dimap
 
+-- | An isomorphism between two functors that are coercible/have the same
+-- internal representation.  Useful for newtype wrappers.
 coercedF :: (forall x. Coercible (f x) (g x), forall x. Coercible (g x) (f x)) => f <~> g
 coercedF = isoF coerce coerce